backend/requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
authentication==1.1.0
autobahn==23.6.2
Automat==22.10.0
bcrypt==4.3.0
billiard==4.2.0
bleach==6.1.0
blinker==1.8.2
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
channels==4.0.0
channels_redis==4.2.1
charset-normalizer==3.3.2
class-registry==2.1.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.0
cryptography==42.0.5
cycler==0.12.1
daphne==3.0.2
diff-match-patch==20230430
distlib==0.3.8
distro==1.9.0
Django==5.0.3
django-allauth==64.1.0
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-crontab==0.7.1
django-filter==24.1
django-fsm==3.0.0
django-import-export==4.1.1
django-js-asset==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
executing==2.0.1
fastapi==0.115.13
filelock==3.13.1
filters==1.3.2
Flask==3.0.3
fonttools==4.53.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
i==2.1.0
idna==3.6
incremental==22.10.0
install==1.3.5
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.10.0
joblib==1.4.2
jwt==1.3.1
kiwisolver==1.4.7
kombu==5.4.0
MarkupSafe==2.1.5
matplotlib==3.9.2
msgpack==1.1.0
npm==0.1.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.91.0
openpyxl==3.1.5
optional-django==0.1.0
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pillow==10.2.0
platformdirs==4.2.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2024.1
redis==5.0.5
regex==2024.4.16
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.5.2
scipy==1.14.1
service-identity==24.1.0
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.4.4
starlette==0.46.2
start==0.2
tablib==3.5.0
threadpoolctl==3.5.0
tqdm==4.67.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.34.3
vine==5.1.0
virtualenv==20.25.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
WTForms==3.1.2
zope.interface==6.2
fastapi
uvicorn
python-dotenv
groq
sentence-transformers
faiss-cpu
pypdf



# Setup Files

## 1. requirements.txt

# Core FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.5.0

# LLM & Embeddings
groq==0.11.0
sentence-transformers==3.1.1

# Vector Store
chromadb==0.5.5

# Graph Processing
networkx==3.3
pyvis==0.3.2

# Document Processing
pypdf==4.3.1
python-docx==1.1.2

# Utilities
python-dotenv==1.0.1
numpy==1.26.4
tiktoken==0.7.0
tenacity==9.0.0

# Testing (optional)
pytest==8.3.0
httpx==0.27.0
